reduce duplication with utils module
